MODIFIED: Code cleanup, removing over 9k unneeded int32(float(uint32(float(2*3))) and adding some float2int32(). Note: when dealing with integer divisions, first cast to float then divide, not vice versa. MODIFIED: OnDamageTaken(), OnTargetCritHit() and OnCritHit() accept int32/uint32 instead of float since damage can be only int32/uint32. MODIFIED: Removed FL2UINT, using float2int32() or (uint32) instead. uint32 casts will be properly cleaned up with another commit.

+++ src/arcemu-world/LocalizationMgr.cpp
@@ -118,7 +118,7 @@ void LocalizationMgr::Shutdown()
 	delete [] m_WorldBroadCast;
 	delete [] m_WorldMapInfo;
 	m_languages.clear();
-	Log.Notice("LocalizationMgr", "Pointer cleanup completed in %.4f seconds.", float(float(getMSTime()-t) / 1000.0f));
+	Log.Notice("LocalizationMgr", "Pointer cleanup completed in %.4f seconds.", (getMSTime()-t) / 1000.0f);
 #undef SAFE_FREE_PTR
 }
