+++ src/arcemu-world/MovementHandler.cpp
@@ -520,7 +520,7 @@ void WorldSession::HandleMovementOpcodes( WorldPacket & recv_data )
 			if( _player->isAlive() && !_player->bInvincible && !_player->GodModeCheat && falldistance > 12 && ( UNIXTIME >= _player->m_fallDisabledUntil ) /*&& movement_info.FallTime > 1000*/ && !_player->m_noFallDamage )
 			{
 				// 1.7% damage for each unit fallen on Z axis over 13
-				uint32 health_loss = float2int32( float( _player->GetUInt32Value( UNIT_FIELD_MAXHEALTH ) * ( ( falldistance - 12 ) * 0.017 ) ) );
+				uint32 health_loss = float2int32( _player->GetUInt32Value( UNIT_FIELD_MAXHEALTH ) * ( falldistance - 12 ) * 0.017f );
 
 				if( health_loss >= _player->GetUInt32Value( UNIT_FIELD_HEALTH ) )
 					health_loss = _player->GetUInt32Value( UNIT_FIELD_HEALTH );
